Switch back to uint256

+++ src/BidStorage.sol
@@ -23,7 +23,7 @@ abstract contract BidStorage {
     /// @param maxPrice The maximum price for the bid
     /// @return bid The created bid
     /// @return bidId The id of the created bid
-    function _createBid(bool exactIn, uint128 amount, address owner, uint256 maxPrice)
+    function _createBid(bool exactIn, uint256 amount, address owner, uint256 maxPrice)
         internal
         returns (Bid memory bid, uint256 bidId)
     {
